+++ src/main/kotlin/org/ionproject/integration/factory/implementations/TimetableFactory.kt
@@ -1,18 +1,18 @@
-package org.ionproject.integration.builder.implementations
+package org.ionproject.integration.factory.implementations
 
 import java.nio.file.Path
-import org.ionproject.integration.builder.interfaces.IGenericFactory
-import org.ionproject.integration.model.internal.generic.AcademicCalendar
+import org.ionproject.integration.factory.interfaces.IGenericFactory
 import org.ionproject.integration.model.internal.generic.IInternalModel
+import org.ionproject.integration.model.internal.generic.Timetable
 import org.ionproject.integration.utils.YamlUtils
 import org.ionproject.integration.utils.orThrow
 
-class AcademicCalendarFactory : IGenericFactory {
+class TimetableFactory : IGenericFactory {
     override fun parse(path: Path): IInternalModel {
         return YamlUtils
             .fromYaml(
                 path.toFile(),
-                AcademicCalendar::class.java
+                Timetable::class.java
             )
             .orThrow()
     }
