+++ src/main/kotlin/org/ionproject/integration/factory/interfaces/IGenericFactory.kt
@@ -1,4 +1,4 @@
-package org.ionproject.integration.builder.interfaces
+package org.ionproject.integration.factory.interfaces
 
 import java.nio.file.Path
 import org.ionproject.integration.model.internal.generic.IInternalModel
