+++ src/test/kotlin/org/ionproject/integration/builder/implementations/ExamScheduleFactoryTest.kt
@@ -3,6 +3,7 @@ package org.ionproject.integration.builder.implementations
 import java.nio.file.Paths
 import java.time.Instant
 import java.util.Date
+import org.ionproject.integration.factory.implementations.ExamScheduleFactory
 import org.ionproject.integration.model.internal.generic.ExamSchedule
 import org.ionproject.integration.utils.YAMLException
 import org.junit.jupiter.api.Assertions.assertEquals
