+++ src/test/kotlin/org/ionproject/integration/builder/implementations/GenericFactoryTest.kt
@@ -1,5 +1,6 @@
 package org.ionproject.integration.builder.implementations
 
+import org.ionproject.integration.factory.implementations.GenericFactory
 import org.ionproject.integration.model.internal.generic.JobType
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
